refactor: split cleanup facades
Move downloader, NLP entity, and segment-city implementation code behind compatibility facades. Keep public imports and test patch seams stable while moving behavior into focused modules.
